requirements.txt
# Main application dependencies
streamlit==1.33.0
streamlit-option-menu==0.3.2

# Data processing and analysis
numpy==1.23.5
pandas==1.5.0
scikit-learn==1.1.2
xgboost==1.6.2
matplotlib==3.6.0
joblib==1.3.2
scipy==1.10.1
statsmodels==0.14.0
imbalanced-learn==0.10.1

# Machine learning and AI
tensorflow==2.15.0
groq==0.4.0

# Computer vision and media processing
opencv-python==4.8.0
mediapipe==0.10.0
pypdf==3.15.1
Pillow==9.5.0

# Google services integration
gspread==5.10.0
oauth2client==4.1.3
google-auth==2.22.0
google-api-python-client==2.86.0

# Communication services
sendgrid==6.10.0
twilio==8.2.0

# Emotion analysis
hume==0.4.0

# Environment and utilities
python-dotenv==1.0.0
watchdog==2.1.9
requests==2.31.0

